feat: mock run 1

+++ arduino/Alex/Alex.ino
@@ -78,7 +78,7 @@ char _recvBuffer, _xmitBuffer;
 
 
 volatile long ultDuration = -1;
-volatile float ultDist = -1;
+volatile long ultDist = -1;
 
 /* --------------------------------------------------------------------------------*/
 // This function needs to be here if not it'll throw an error
